backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.0

# Phase 3: Storage & AI
boto3>=1.34.46
cloudflare>=4.3.1
anthropic>=0.58.2
openai>=1.97.0

# Background Processing & Caching
redis>=5.2.1
celery[redis]>=5.5.3
flower>=2.0.1

# Additional Dependencies
httpx>=0.28.1
pillow>=10.1.0
requests>=2.31.0
aiosqlite>=0.19.0
asyncpg>=0.29.0